Move MockAG95
